create a API class for working with residents
