--- SocialCareCaseViewerApi/V1/UseCase/Interfaces/IGetAllUseCase.cs
@@ -1,10 +1,0 @@
-using SocialCareCaseViewerApi.V1.Boundary.Requests;
-using SocialCareCaseViewerApi.V1.Boundary.Response;
-
-namespace SocialCareCaseViewerApi.V1.UseCase.Interfaces
-{
-    public interface IGetAllUseCase
-    {
-        ResidentInformationList Execute(ResidentQueryParam rqp, int cursor, int limit);
-    }
-}
